+++ extensions/slack/src/shared/client/WebClient.ts
@@ -1,5 +1,13 @@
 import { getPreferenceValues } from "@raycast/api";
 import { WebClient } from "@slack/web-api";
 
+export interface SlackConversation {
+  id?: string;
+  name?: string;
+  shared_team_ids?: string[];
+  internal_team_ids?: string[];
+  is_private?: boolean;
+}
+
 const accessToken = getPreferenceValues<{ accessToken: string }>().accessToken;
 export const slackWebClient: WebClient = new WebClient(accessToken);